requirements.txt
Django==1.8.2
Pillow==3.1.1
argparse==1.2.1
django-appconf==1.0.1
django-bootstrap3==5.4.0
django-braces==1.8.0
django-compressor==1.5
django-debug-toolbar==1.3.2
django-friendship==1.2.0
django-image-cropping==1.0.1
django-user-accounts==1.1
easy-thumbnails==2.2
pytz==2015.4
requests==2.9.1
six==1.9.0
sqlparse==0.1.15
wheel==0.24.0
wsgiref==0.1.2
